+++ packages/polaris-viz/src/components/DonutChart/Chart.tsx
@@ -21,9 +21,10 @@ import {
   useColorVisionEvents,
   useWatchColorVisionEvents,
 } from '../../hooks';
+import {Arc} from '../Arc';
 
 import styles from './DonutChart.scss';
-import {Arc, InnerValue} from './components';
+import {InnerValue} from './components';
 
 const FULL_CIRCLE = Math.PI * 2;