+++ packages/polaris-viz/src/components/LineChart/LineChart.tsx
@@ -120,7 +120,12 @@ export function LineChart(props: LineChartProps) {
       )}
       <ChartContainer theme={theme}>
         {state !== ChartState.Success ? (
-          <ChartSkeleton state={state} errorText={errorText} theme={theme} />
+          <ChartSkeleton
+            type="Default"
+            state={state}
+            errorText={errorText}
+            theme={theme}
+          />
         ) : (
           <Chart
             data={dataWithDefaults}