+++ packages/polaris-viz/src/constants.ts
@@ -71,3 +71,4 @@ export const LEGENDS_TOP_MARGIN = 16;
 export const PREVIEW_ICON_SIZE = 12;
 export const BAR_CONTAINER_TEXT_HEIGHT = 48;
 export const ANNOTATIONS_LABELS_OFFSET = 10;
+export const ARC_PAD_ANGLE = 0.02;